A workaround to fix long boot animation issue.
A workaround in PermissionController to fix systemui crash issue because of lack of role permissions, resulting in long time of turning off the boot animation. This bug happens if rebooting in first boot while runtime-permissions.xml has not been persisted. There is a delay before persistence of runtime-permissions.xml in PermissionManagerService.

Test: manual

Change-Id: I14f76b6e02108a3fea81f8f83b28c4fa5ca0b626