Dismiss permission dialog with the Esc key
Dialogs and activities with "close when touch outside" will
automatically close when the user presses the Esc key; however,
this is currently not the case with permission dialogs. This is
because these dialog are handling the outside touch by themselves.

The solution I'm proposing is to handle the Esc key the same way.

Bug: 315954941
Test: Google Maps location grant; Esc will close the dialog.
Change-Id: I390845198382ce4821c74a22c948da65d37aaef8