Opt-in SafetyCenterActivity for predictive back-gestures.
https://developer.android.com/guide/navigation/custom-back/predictive-back-gesture#opt-predictive

As per b/303779019#comment6. Only doing this for SafetyCenterActivity rather than doing this at the app-level as this behavior may not be desirable for other activities owned by PermissionController (although we may want to revisit this or enable this for other Settings related activities).

Bug: 303779019
Change-Id: If92fd557d0a8bcda8958777ff789f59f85179d63
Relnote: Opt-in SafetyCenterActivity for predictive back-gestures
Test: manual