Generalise the code for various profile types - 1
This code generalises the code for creating safety source entries for
any profile type. Currently the code handles the primary profile and
the managed profiles specifically.

Don't include the private profile when in quiet mode
If the private profile is in quiet mode then we don't want to show
anything related to it in the safety center. Therefore,
UserProfileGroup should not include it in such cases at all.

Bug: 286539356
Test: manual
Change-Id: I77be00f140a5c7be74caa2dddf673a81562851ef