[ECM] Enforce MANAGE_ENHANCED_CONFIRMATION_STATES
Enforce the MANAGE_ENHANCED_CONFIRMATION_STATES permission in
EnhancedConfirmationService instead of MANAGE_APPOPS.

Previously, we enforced MANAGE_APPOPS. But enforcing this permission is
a leaky abstraction: callers should not need to even know we're dealing
with appops in the service. Also, this permission is too powerful: we
don't want to ask callers to use this permission. It's also hidden API.

Enforcing a new permission specific to ECM solves all of this.

Bug: 320512579
Test: atest EnhancedConfirmationManagerTest
Change-Id: I992bcff2689bc66c00d17a721ea441c70ca2d38d