[ECM] Make setClearRestrictionAllowed SystemApi
We plan to call EnhancedConfirmationManager::setClearRestrictionAllowed
a hidden API method, from PermissionController. But,
PermissionController can't access hidden API methods. In order to make
this possible, setClearRestrictionAllowed will need to be upgraded from
hidden API to SystemApi.

The reason we need to call setClearRestrictionAllowed from
PermissionController is that we plan to move the
ActionDisabledByAppOpsDialog activity into PermissionController, and
modify it to call setClearRestrictionAllowed instead of directly setting
appops.

Bug: 322026141
Fix: 320517290
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
Change-Id: Ibd068ac456846fd1af8e867a6b336853a7835c4a