[ECM] Use unique requestCodes for PendingIntent
Use a unique requestCode for each call to PendingIntent.getActivity.

Without this, each call returns the same object, even if the supplied *extra*s differ. This is a problem because, it would mean that after the first call to PendingIntent.getActivity, each subsequent call will have the wrong extras.

Fix: 322891228
Test: manual
Change-Id: If774bfc1e03963ec32980ea7c935774eb930eaf0