[DON'T BLOCK] Test ownership migration rules
This CL is created as a best effort to migrate test targets
to the new android ownership model. If you find incorrect or unnecessary
attribution in this CL, please create a separate CL to fix that.

For more details please refer to the link below,
<add g3 doc link>

Bug: 304529413
Test: N/A
Change-Id: Icbe839de65c4449f764bd579a9f8c43f9fac770d
Merged-In: Icbe839de65c4449f764bd579a9f8c43f9fac770d