Remove test for revoking permission that's not requested.
A pervious attempt to fix these tests was at ag/26049635, where we
removed the part of these test that asserts an exception is thrown.
However, because we are on main the test is still running for U QPR as
well so we are now getting an error for the old behavior.

Since we no longer throw an exception right now, there's actually no
visible behavior to test against, and we can simply remove the test to
resolve the discrepancy.

Fixes: 323290220
Test: presubmit
Change-Id: I1b1aea268baf803b3e0fa76f8c0dfd9c76071eab