[ECM] Define ACTION_SHOW_ECM_RESTRICTED_SETTING_DIALOG
We're planning to move the "Restricted setting" dialog into
PermissionController. But, this dialog's intent action string,
Settings.ACTION_SHOW_RESTRICTED_SETTING_DIALOG, is currently bound to
the Settings app.

So, to ease the transition, this change defines a new intent action
string.

This string should only need to be accessed by PermissionController,
which handles the intent, and EnhancedConfirmationManager, which
constructs the intent.

Bug: 322026141
Test: manual
Change-Id: I1ba1d5bc778dd39239abb0dc6e0f0f32bea0c2b2