Renamed PermissionState to reuse in SystemAPI
The class `PermissionState` is being used in ag/26066306 as SystemAPI. Hence renaming this to an unused class, which will be cleaned up in a follow up CL. Effort to rename as well as recuse in single CL has failed in pre-submit with the error `FAILED: out/soong/hiddenapi/hiddenapi-stub-flags.txt` and `Class duplication: Landroid/permission/PermissionState`. Hence, created this CL separately to rename first.

Bug: 322876542
Test: manual
Change-Id: I44dcf2c8a2b0ba1fd9f9e9e4492f1bc4d7566b21