[ECM] Point intent at new EnhancedConfirmationDialog
We recently moved---well, copied---the "Restricted setting" dialog into
PermissionController. But, we're not using the new dialog yet.

The new dialog receives a different intent action than the old. So, to
migrate to the new dialog, we'll need to update all code that launches
the old intent. Fortunately, we've encapsulated this in
EnhancedConfirmationManager, so this is a one-line change.

Bug: 322026141
Test: CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
Change-Id: Ied3e2712e409e056945be08e5911ae232149de68