[ECM] Use enforceCallingOrSelfPermission
EnhancedConfirmationService enforces the
MANAGE_ENHANCED_CONFIRMATION_STATES permission by calling
enforceCallingPermission.

But, this service is sometimes called from SystemServer via non-IPC.
In practice, this results in a SecurityException.

This exception is documented behavior of enforceCallingPermission,
which says "if you are not currently processing an IPC, this function
will always throw the SecurityException".

The solution used here is to instead call
enforceCallingOrSelfPermission.

Fix: 325485150
Test: Test: atest CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
Change-Id: I45057f00701b746013e9be4213df71b09d24b195