Create Role.get(Application)RestrictionIntent().
So that restriction is now an official concept in role and can be
applied to all role UI surfaces in a consistent way.

The change for request role dialog will need to be in a later CL.

LOW_COVERAGE_REASON=Refactor that removes more code than added, and
touches auto and wear code.

Bug: 310218979
Test: RoleManagerTest
Change-Id: Ibf94928ec7380156c2dd7b4e7097e55fa8561533