Properly handle exceptions for IRoleController
Update the methods in IRoleController to catch the exception and
send it back to the caller's site.

This CL is a general fix for exception handling in IRoleController. It
also fixes and an issue we previously encountered where the role migration
failed when PC version mismatched. More details:

In the very rare cases that the PermissionController is not in an
updated version, it's possible that the system-server side of the role
migration code is in place but PC doesn't have the code to return the
legacyFallbackEnabledRoles. In this case currently we throw
UnsupportedOperationException. However, after this exception is thrown,
we didn't resolve the result for the cross-binder remote callback that's
being sent to PC. Hence in system-server it will wait until this call is
timedout.

Fixes: 325264710
Test: RoleManagerTest and additionally tested locally
LOW_COVERAGE_REASON=b/330904893

Change-Id: If77b32e39db5b0853cd54c8b74c871b0684611f5