[ECM] Allow opt-in to installer allowlist
This allows the ability to opt-in to the "ECM trusted installer"
feature.

A "trusted installer" is any package which is declared in an
<enhanced-confirmation-trusted-installer> element in the
/data/etc/enhanced-confirmation.xml file.

If at least one trusted installer is declared on the device, this
feature is considered opted in. Otherwise, the feature is considered
opted out.

If the feature is opted in, then:

- For packages which were installed by an untrusted installer (an
  installer that is *not* declared as a trusted installer): ECM will
  treat these packages as restricted; i.e., ECM will block these
  packages from accessing restricted settings.
- For packages which were installed by a trusted installer: ECM will
  only treat these packages as restricted if the installer marks them
  as untrustworthy (i.e., LOCAL/DOWNLOADED).

If the feature is opted out, then:

- ECM will behave as if all packages are trusted installers. That is:
  ECM will only treat a package as restricted if its installer marks
  it as untrustworthy (i.e., LOCAL/DOWNLOADED).

Bug: 327469700
Test: atest CtsPermissionUiTestCases:android.permissionui.cts.EnhancedConfirmationManagerTest
Change-Id: I6e2036664b21b4ab546bc5bf94002ec7cf134043
LOW_COVERAGE_REASON=(1) each line should already be covered by EnhancedConfirmationManagerTest and (2) allowlist is hardcoded into system image so is hard to test every condition e2e