[ECM] Fix package name bug
We were accidentally passing the wrong package name to
setClearRestrictionAllowed(). This resulted it in the AppInfo page never
actually allowing the user to clear app restrictions if only a runtime
permission was restricted.

"getPackageName()" gets the package name of PermissionController itself.

"getCallingPackage()" gets the package name of the package which invoked
the permission request.

Fix: 328535805
Test: atest EnhancedConfirmationManagerTest
Change-Id: I575026176269e1bb26c6083a426537d0cf8de0a0