[ECM] Suppress repeat permission ECM dialogs
Historically, when a permission is denied multiple times by the user,
we suppress requests for that permission.

However, the ECM dialog currently shows *every* time the app requests
it. This is undesirable, because it isn't in the spirit of the
above-stated behavior; it might be prone to abuse.

This change modifies GrantPermissionsActivity so that, if the user has
already seen the ECM dialog for this app *ever*---even outside of
GrantPermissionsActivity---then GrantPermissionsActivity won't show
the ECM dialog for this app again. (With one small exception: if the
app requests multiple permission groups in the same request, we'll
show multiple ECM dialogs for that request.)

We aren't too worried about users missing out on information here,
because from the app's perspective, the permission was denied; the app
should be guiding the user to try manually enabling the permission,
which will result in the user seeing the dialog again.

(This change also fixes a bug: we were using getCallingPackage(). But,
we should use mTargetPackage, which may differ if System Server
requested this for another app.)

Fix: 328656482
Test: atest EnhancedConfirmationTest
Change-Id: I090cc6d980343f19ede163dc526f7e518086b0de