Supporting external device permission change in App Permissions
ScreenRecording: https://screenshot.googleplex.com/4qT3QHMEkBk5WRS.png

1) Allows users to update the external device permissions
2) Added CTS tests to verify the permission changes
3) Naming change from VirtualDevice to ExternalDevice

Bug: 322876988
Test: atest CtsPermissionMultiDeviceTestCases:android.permissionmultidevice.cts.AppPermissionsTest

Change-Id: Ibf00439826526a61a42346e8ed385e80cc16dba8

LOW_COVERAGE_REASON=New tests added in this CL provide 90% incremental coverage. Since the test module is new, it cannot be submitted in presubmit as is, needs to run for a while in postsubmit to meet SLO criteria. Since the test module is now only part of postsubmit, coverage is shown as low. There will be another followup CL in a week to update the module to run as part of presubmit