Restrict packages with install source PACKAGE_SOURCE_UNSPECIFIED and targetSdkVersion is V+.
An installer app can create a install session using the PackageInstaller
APIs and session params and commit it. Without setting the source, the
install source will default to UNSPECIFIED. The installingPackageName is
the installer app packageName.

The only time originatingPackageName is populated is when the installer
app opens PackageInstallerAcitvity and installed an app. This will set
the installingPackageName to PIA instead of the installer app and the
packageSource will be set to LOCAL_FILE or DOWNLOADED_FILE.

Other than that, we should only check installingPackageName for the
installer app/parent app.

Bug: 329441688
Test: atest EnhancedConfirmationManagerTest
Test: atest AppPermissionTest

LOW_COVERAGE_REASON=b/329185137

Change-Id: Iac0ddbe984f191e947ab2659f7fe2def5cef1931