[ECM] Log RESULT_SUPPRESSED and RESULT_CANCELLED
1. When an ECM dialog is suppressed (i.e., we prevent it from being
shown more than once), log it as a "dialog result" with the "result"
type being "SUPPRESSED".

2. When the user "swipes up" out of the app, log this as CANCELLED.
(Right now we're not logging this event if the dialog is swiped out:
onDestroy will be invoked despite onDialogResult never having been
called. We can fix this by defaulting to CANCELLED and moving our
log into onDestroy.)

Bug: 324254847
Test: manual
Change-Id: I66d6697376e6e333120fb3d5c522c324519dd99f