New UI requests for Restriction Dialog
Add a `EXTRA_IS_ECM_DEFAULT` boolean to the restriction intent and
generate dialog with different title and description strings depending
on the boolean.

The boolean will be set to true as default, and false when the dialog is triggered in-app/Default apps or in-app/Runtime permission.

Bug: 331268773
Test: atest RoleManagerTest
Test: atest AppPermissionTest

LOW_COVERAGE_REASON=UX string changes

Change-Id: Ic9b3b754febc727018725e897c95f49f79dc38bd