Switched to long date format in Permission Usage Details screen
Bug: 323767307
Test: Manual + pre-submit

After consulting with the team and TPM, we decided to use the long date format. This will improve readability and ensure voice assistants pronounce 'twenty twenty-four' correctly.

Change-Id: I246c3092860de543225cd560d0575c6dcaf189b6

LOW_COVERAGE_REASON=This is a 1-line change and does not alter the existing coverage. It is possible that there are no tests running in presubmit for this module, which would explain why the incremental coverage is shown as 0.