Fix disableLocation bug in PermissionTest30WithBluetooth
In our @Before method, we check whether Location is enabled, setting
locationWasEnabled.

In our @After method, if locationWasEnabled==false, we disable Location.

But, if @Before throws an exception, then locationWasEnabled currently
defaults to false, even if we never actually checked the location.
@After is still run, and it sees locationWasEnabled==false, so it
disables location.

To mitigate this, default locationWasEnabled to true.

(Credit to meghana.manjunatha@amd.com for reporting this issue and
for offering fix suggestions.)

Bug: 328470600
Test: run cts -m CtsPermissionUiTestCases -t  android.permission3.cts.PermissionTest30WithBluetooth
Change-Id: Ic8b45706c51f9fd1f386ee9deba1951a336170d3