Fix auto revoke logic for split permissions
Lets say permissoin A splits into new permission B,
and permission B's group is eligible for revocation,
then we should exempt B's permission group revocation
(based on A's split), only if A's permission group
is granted.

If A's permission group is not granted, and B's
permission group is eligible for revocation, we
should revoke it.

Also, fixing the target sdk check.

Fix: 339306269
Test: atest AutoRevokeTest
Change-Id: I7316613ce9160d64a5a0db63fb9483c41c3a5f98