Stronger scroll search in BaseUsePermissionTest
This is a CP of https://android-review.git.corp.google.com/c/platform/cts/+/3080510

scrollTextIntoView did not work well for Wear devices that have vertical
global surfaces, because the method would scroll to those surfaces and
give up if it cannot find the text. We are now adding a check with the
UiAutomatorUtils2 API that is more robust in handling such cases.

Bug: 339594637
Bug: 340146992
Bug: 332620051
Change-Id: Ic5026bb5c3e7521360719cc4790fa1e6cda6653a
Test: android.permission3.cts.PermissionSplitTest#testBodySensorSplit32