Fix clickPermissionRequestSettingsLink for Wear
This is a CP of aosp/3084808

It failed to find deep link if it isn't the screen,
since it tried to find the deep link on the bottom of the screen.

Make the deep link visible in the screen by moving the detail message
on the top of the screen.

Bug: 332620051
Test: android.permission3.cts.PermissionTest29
Change-Id: I5d38e6e829a6c66ebbc13206988a0895d56aba7a