[CTS] Don't call waitFindObjectOrNull on each request
This clickNotificationPermissionRequestAllowButtonIfAvailable method
is called after upon every permission request, closing the notification
dialog if it is present.

But, to determine whether that dialog is present, it uses
waitFindObjectOrNull, which has the unintentional side effect of trying
to scroll the screen, regardless of which screen we're on.

The result is that each time we launch a permission request, we
immediately start swiping the screen. This is suspected to cause
occasional test flakes, because it might accidentally dismiss the grant
dialog.

To solve this, instead of using waitFindObjectOrNull, just use the raw
UiAutomator wait(...) feature, which doesn't automatically try to
scroll.

Also, the notification dialog only appears for targetSdk<T apps.
So, don't bother if the app is T+.

Change-Id: I30d543aef452cfe3d4077607d707a2ba1828f3bf
Bug: 340954765
Test: atest CtsPermissionUiTestCases
LOW_COVERAGE_REASON=TEST_ONLY