[CTS] Don't call waitFindObjectOrNull on targetSdkWarning
On pre-Android-V devices, the clearTargetSdkWarning method is
called after each time our test installs a package. Its purpose
is to close the DeprecatedTargetSdkVersionDialog, if shown.

This change fixes two issues with this code.

1)

To determine whether that dialog is present, we currently use
waitFindObjectOrNull, which has the unintentional side effect of trying
to scroll the screen, regardless of which screen we're on.

The result is that each time we install a package, we
immediately start swiping the screen. This is suspected to cause
occasional test flakes.

To solve this, instead of using waitFindObjectOrNull, just use the raw
UiAutomator wait(...) feature, which doesn't automatically try to
scroll.

2)

We currently don't confirm that the on-screen dialog is indeed the
DeprecatedTargetSdkVersionDialog before closing it. If any other dialog
is shown, we'll click any visible dialog's "button1". Fix that by first
waiting for something unique to this dialog.

Bug: 339536486
Test: atest CtsPermissionUiTestCases with debug.wm.disable_deprecated_target_sdk_dialog not set
Change-Id: I93ee37a2f07b9a10802b5c757d80ee803a259fd3
LOW_COVERAGE_REASON=TEST_ONLY