Make GrantPermissionsAutoViewHandler AlertDialog scrollable
The AlertDialog shown by GrantPermissionsAutoViewHandler will crop the
buttons when the title and subtile are too long for the dialog. This
change makes the dialog scrollable with the message. Also, fix test so
it scrolls to the bottom of the dialog for deny buttons.

Bug 317442167

Test: atest android.permissionui.cts.PermissionTest29#testDenyBackgroundWithPrejudice
LOW_COVERAGE_REASON=Auto code

Change-Id: I2c1f740217079d14c95ff88209a8da74731fa790